refactor player movement and fix strafing bug
