+++ src/game/player/movement/constants.js
@@ -1,0 +1,16 @@
+export const walkSpeed = 25; // Units per second for walking
+export const runSpeed = 50;  // Units per second for running
+export const flySpeed = 60;  // Flight speed
+export const GRAVITY = 140;
+export const JUMP_FORCE = 70;
+
+// Dynamic ground parameters
+export const BASE_GROUND_Y = 0;
+
+// Central wall parameters (must match objects/index.js)
+export const WALL_RADIUS = 960;
+export const WALL_THICKNESS = 5;
+export const WALL_HEIGHT = 30;
+
+// Player collision radius on XZ plane
+export const PLAYER_RADIUS = 2;
